Created classes to define 1D slab
Classes were created or changed to allow the creation of a 1D slab populated
with Cell objects. Angular fluxes were set up by creating an AngleDependent
class which rely on a Quadrature object for the cosines and weights.

- Created Slab class to contain a const Settings and const Layout object

- Created AngleDependent class to represent angular fluxes in cell objects

- Created Quadrature class to represent cosines and weights, defaulting to
Legendre-Gauss quadrature with N=64

- Modified Cell class to contain midpoint and outgoing angular flux

- Modified Segment class to contain initial guess of scalar flux

- Modified Layout class to initialize cells with guess of initial scalar flux
from the segment they belong to

- Modified Settings class to contain initial guess of k eigenvalue and
qudrature to be used for ordinates