Created classes to define AngularFlux
Angular flux is represented by a map of energy group keys to AngleDependent
objects. The quadrature class was removed in place of a separate file
(quadrature.dsv) containing cosines and weights that would be read on the
instantiation of a AngleDependent object.

- Modified AngleDependent class constructor to only use one input parameter.
Cosines and weights now set by including quadrature.dsv in initialization.
Added overload of operator<<().

- Created AngularFlux class to represent group-dependent angular fluxes.

- Modified Cell class to use AngularFlux object instead of AngleDependent
object.

- Deleted Quadrature class.

- Removed Settings class references to Quadrature object. Added convergence
tolerances for k eigenvalue and sclalar flux. Added initial guess of k
eigenvalue.

- Modified sn-solver to reflect changes in Settings class.