Fixed bug in neutron transport methods
Main sources of the problem were normalization of the scalar flux and
calculation of the fission source term. It remains uncertain if the source term
in the sweep methods should be normalized by the sum of the quadrature weights
(2.0).

- Modified Slab class with methods for solving fixed source problems and
printing angular fluxes.

- Modified Cell class with methods for handling external sources. Methods to
update the fission source were fixed. Methods added for printing angular flux.

- Modified AngleDependent class to normalize weighted sum.

- Modified AngularFlux class to return reference to underlying data.

- Modified GroupDependnet class with redefined vector inner product method and
new element-wise product method.